refactor(connect-common): promote MethodPermission and MethodInfo to public API
Move MethodPermission and MethodInfo from packages/connect/src/core/AbstractMethod.ts
into packages/connect-common/src/types/method.ts and re-export them through the
@trezor/connect-common types barrel. They now appear in the @trezor/connect,
@trezor/connect-web and @trezor/connect-webextension public surface and no
longer require deep-path imports such as
'@trezor/connect/src/core/AbstractMethod'.

AbstractMethod re-exports them for internal connect modules so the file-local
import path keeps working. Consumers that previously used the deep import
(connectPopupThunks, connectPopupTypes, ethereumSignTransaction method hook,
suite-native PermissionConfirmation) now import from @trezor/connect.

Refs #27374
