+++ packages/connect-common/src/types/index.ts
@@ -4,6 +4,7 @@ export * from './coinInfo';
 export * from './device';
 export * from './fees';
 export type * from './firmware';
+export type * from './method';
 export * from './params';
 export * from './settings';
