+++ packages/connect/src/core/AbstractMethod.ts
@@ -6,6 +6,8 @@ import type {
     CoreEventMessage,
     DeviceState,
     FirmwareCapability,
+    MethodInfo,
+    MethodPermission,
     PrecomposeResultFinal,
     StaticSessionId,
     UiRequestButtonData,
@@ -19,28 +21,19 @@ import type { Device } from '../device/Device';
 import type { UiPromiseCreator } from '../events/ui-promise';
 
 export { DEFAULT_FIRMWARE_RANGE };
+// TODO: drop this re-export and migrate the ~50 internal `packages/connect/src/api/**`
+// imports to pull `MethodPermission` / `MethodInfo` directly from `@trezor/connect-common`.
+// Kept here as a transitional shim to keep this PR focused on the public-surface change;
+// the cleanup will follow up alongside the deep-import ESLint guardrail (#27376).
+export type { MethodInfo, MethodPermission };
 
 export type Payload<M> = Extract<CallMethodPayload, { method: M }> & { override?: boolean };
 export type MethodReturnType<M extends CallMethodPayload['method']> = CallMethodResponse<M>;
 
-export type MethodPermission = 'read' | 'write' | 'management' | 'push_tx';
 export type DeviceMode =
     | typeof UI_REQUEST.SEEDLESS
     | typeof UI_REQUEST.BOOTLOADER
     | typeof UI_REQUEST.INITIALIZE;
-
-export type MethodInfo = {
-    // static fields
-    useUi: boolean;
-    useDevice: boolean;
-    useDeviceState: boolean;
-    name: string;
-    requiredPermissions: MethodPermission[];
-    // available after init
-    info: string;
-    precomposed?: PrecomposeResultFinal;
-    confirmation?: UiRequestConfirmation['payload'];
-};
 
 export type MethodContext = {
     sendCoreMessage: (message: CoreEventMessage) => void;
