+++ suite-common/connect-popup/src/connectPopupThunks.ts
@@ -9,8 +9,9 @@ import TrezorConnect, {
     type CallMethodKeys,
     type CallMethodParams,
     type CallMethodPayload,
+    type MethodInfo,
+    type MethodPermission,
 } from '@trezor/connect';
-import { type MethodInfo, type MethodPermission } from '@trezor/connect/src/core/AbstractMethod';
 import { connectCallableMethods } from '@trezor/connect-common';
 import { TypedError, serializeError } from '@trezor/connect-common/src/constants/errors';
 import { DEEPLINK_VERSION } from '@trezor/connect-common/src/data/version';
